requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-login==0.6.3
werkzeug>=3.1.0
moviepy==2.1.2
openai-whisper==20240930
transformers==4.51.3
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
librosa==0.10.1
numpy>=1.25.0
gTTS==2.5.4
soundfile==0.13.1
scipy>=1.11.0
scikit-learn>=1.3.0
face_alignment==1.3.5
imageio==2.31.1
pysubs2==1.4.2
typer>=0.9.0
gradio==4.19.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
python-magic==0.4.27
wav2lip==1.0.0
pytest>=7.4.0
tensorflow-gpu>=2.12.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
pydantic>=2.0.0
loguru>=0.7.0
